+++ valid8/tests/test_entry_points_inline.py
@@ -2,19 +2,19 @@ import pytest
 from valid8 import ValidationError
 
 
-def test_quick_valid():
-    """ Tests the quick_valid function """
+def test_validate_():
+    """ Tests the validate function """
 
-    from valid8 import quick_valid
+    from valid8 import validate
 
     # nominal
     surf = 2
-    quick_valid('surface', surf, instance_of=int, min_value=0)
+    validate('surface', surf, instance_of=int, min_value=0)
 
     # error wrong value
     surf = -1
     with pytest.raises(ValidationError) as exc_info:
-        quick_valid('surface', surf, instance_of=int, min_value=0)
+        validate('surface', surf, instance_of=int, min_value=0)
     e = exc_info.value
     assert str(e) == "Error validating [surface=-1]. " \
                      "TooSmall: x >= 0 does not hold for x=-1. Wrong value: [-1]."
@@ -22,24 +22,24 @@ def test_quick_valid():
     # error wrong type
     surf = 1j
     with pytest.raises(ValidationError) as exc_info:
-        quick_valid('surface', surf, instance_of=int, min_value=0)
+        validate('surface', surf, instance_of=int, min_value=0)
     e = exc_info.value
     assert str(e) == "Error validating [surface=1j]. " \
                      "HasWrongType: Value should be an instance of <class 'int'>. Wrong value: [1j]."
 
 
-def test_readme_usage_quick_valid_customization():
-    """ Tests the various customization options for quick_valid """
+def test_readme_usage_validate__customization():
+    """ Tests the various customization options for validate """
 
-    from valid8 import quick_valid
+    from valid8 import validate
     from math import isfinite
 
     surf = 1j
 
     # (A) custom error message (exception is still a ValidationError)
     with pytest.raises(ValidationError) as exc_info:
-        quick_valid('surface', surf, instance_of=int, min_value=0,
-                    help_msg="Surface should be a positive integer")
+        validate('surface', surf, instance_of=int, min_value=0,
+                 help_msg="Surface should be a positive integer")
     e = exc_info.value
     assert str(e) == "Surface should be a positive integer. Error validating [surface=1j]. " \
                      "HasWrongType: Value should be an instance of <class 'int'>. Wrong value: [1j]."
@@ -49,7 +49,7 @@ def test_readme_usage_quick_valid_customization():
         help_msg = 'Surface should be a positive integer'
 
     with pytest.raises(ValidationError) as exc_info:
-        quick_valid('surface', surf, instance_of=int, min_value=0, error_type=InvalidSurface)
+        validate('surface', surf, instance_of=int, min_value=0, error_type=InvalidSurface)
     e = exc_info.value
     assert isinstance(e, InvalidSurface)
 
@@ -58,8 +58,8 @@ def test_readme_usage_quick_valid_customization():
         help_msg = 'Surface should be > {minimum}, found {var_value}'
 
     with pytest.raises(ValidationError) as exc_info:
-        quick_valid('surface', surf, instance_of=int, min_value=0,
-                    error_type=InvalidSurface, minimum=0)
+        validate('surface', surf, instance_of=int, min_value=0,
+                 error_type=InvalidSurface, minimum=0)
     e = exc_info.value
     assert isinstance(e, InvalidSurface)
     assert str(e) == "Surface should be > 0, found 1j. Error validating [surface=1j]. " \
@@ -67,7 +67,7 @@ def test_readme_usage_quick_valid_customization():
 
 
 def test_wrap_valid():
-    """ Tests the wrap_valid context manager """
+    """ Tests the validation context manager """
 
     from valid8 import wrap_valid
     from math import isfinite